Misc optimizations

+++ src/main/base/llm_base_module.py
@@ -1,11 +1,11 @@
-from utils.tokenizer_utils import add_agent_tokens, add_additional_tokens
+from exception.exceptions import TuningModuleFunctionException
 
 from arguments.arguments import TuneArguments, MergeArguments, PushArguments
 
 from peft import LoraConfig, get_peft_model, prepare_model_for_kbit_training, TaskType, AutoPeftModelForCausalLM, PeftModel
-from trl import SFTTrainer, SFTConfig, setup_chat_format
+from trl import SFTTrainer, SFTConfig
 from transformers.trainer_utils import get_last_checkpoint
-from utils.model_utils import get_all_layers, get_all_linear_layers
+from utils.model_utils import get_all_layers, get_all_linear_layers, prepare_model_vocabulary
 from utils.dataset_utils import load_dataset
 import os
 import shutil
@@ -15,19 +15,18 @@ import shutil
 
 
 def fine_tune_base(arguments: TuneArguments, tokenizer, base_model) -> None:
-    if arguments.additional_vocabulary_tokens is not None:
-        add_additional_tokens(tokenizer, base_model, arguments.additional_vocabulary_tokens)
-    if arguments.use_agent_tokens:
-        add_agent_tokens(tokenizer, base_model)
-    if arguments.is_chat_model or (arguments.train_file.endswith(".jsonl")):
-        base_model, tokenizer = setup_chat_format(base_model, tokenizer)
     print(f"Starting fine-tuning of base model {arguments.base_model} for {arguments.new_model}")
     print('')
-    output_dir = f"{arguments.output_directory}/checkpoints/{arguments.new_model}"
-    lora_dir = f"{arguments.output_directory}/adapters/{arguments.new_model}"
+    output_dir = f"{arguments.output_directory}{os.sep}checkpoints{os.sep}{arguments.new_model}"
+    lora_dir = f"{arguments.output_directory}{os.sep}adapters{os.sep}{arguments.new_model}"
     if not arguments.no_checkpoint:
         print(f'Checkpointing to {output_dir}')
         print('')
+
+    if os.path.exists(lora_dir) and not arguments.overwrite_output:
+        raise TuningModuleFunctionException(f'cannot overwrite existing LoRA directory({lora_dir}) when `--overwrite-output` CLI argument is not set to "true"', 'FINE_TUNE')
+
+    base_model, tokenizer = prepare_model_vocabulary(arguments, base_model, tokenizer)
 
     ds = load_dataset(arguments)
 
@@ -62,7 +61,7 @@ def fine_tune_base(arguments: TuneArguments, tokenizer, base_model) -> None:
         per_device_train_batch_size=arguments.batch_size,
         per_device_eval_batch_size=arguments.batch_size,
         gradient_accumulation_steps=arguments.gradient_accumulation_steps,
-        overwrite_output_dir=True,
+        overwrite_output_dir=arguments.overwrite_output,
         optim=arguments.optimizer_type,
         save_strategy=arguments.save_strategy,
         save_steps=arguments.save_steps,
@@ -84,7 +83,7 @@ def fine_tune_base(arguments: TuneArguments, tokenizer, base_model) -> None:
         do_eval=arguments.do_eval,
         # TODO - is this ignored bt SFTTrainer?
         max_seq_length=arguments.max_seq_length,
-        neftune_noise_alpha=5.0 if arguments.is_instruct_model else None,
+        neftune_noise_alpha=arguments.neftune_noise_alpha if arguments.is_instruct_model else None,
         dataset_text_field="text" if not arguments.train_file.endswith("jsonl") else None,
         use_ipex=arguments.cpu_only_tuning
     )
@@ -108,7 +107,7 @@ def fine_tune_base(arguments: TuneArguments, tokenizer, base_model) -> None:
 
     print('')
     print(f'Saving LoRA adapter to {lora_dir}')
-    if os.path.exists(lora_dir):
+    if os.path.exists(lora_dir) and arguments.overwrite_output:
         shutil.rmtree(lora_dir)
 
     train.model.save_pretrained(lora_dir)
@@ -120,15 +119,17 @@ def fine_tune_base(arguments: TuneArguments, tokenizer, base_model) -> None:
 
 
 def merge_base(arguments: MergeArguments, tokenizer, base_model, bnb_config) -> None:
-    if arguments.additional_vocabulary_tokens is not None:
-        add_additional_tokens(tokenizer, base_model, arguments.additional_vocabulary_tokens)
-    if arguments.use_agent_tokens:
-        add_agent_tokens(tokenizer, base_model)
-    if arguments.is_chat_model:
-        base_model, tokenizer = setup_chat_format(base_model, tokenizer)
-    lora_dir = f"{arguments.output_dir}/adapters/{arguments.new_model}"
-    model_dir = f'{arguments.output_dir}/merged-models/{arguments.new_model}'
+    lora_dir = f"{arguments.output_dir}{os.sep}adapters{os.sep}{arguments.new_model}"
+    model_dir = f'{arguments.output_dir}{os.sep}merged-models{os.sep}{arguments.new_model}'
     print(f"merging {arguments.base_model} with LoRA into {arguments.new_model}")
+
+    if not os.path.exists(lora_dir):
+        raise TuningModuleFunctionException(f'cannot merge model because LoRA adapter @ {lora_dir} is missing', 'MERGE')
+
+    if os.path.exists(model_dir) and not arguments.overwrite_output:
+        raise TuningModuleFunctionException(f'cannot overwrite existing model directory({model_dir}) when `--overwrite-output` CLI argument is not set to "true"', 'MERGE')
+
+    prepare_model_vocabulary(arguments, base_model, tokenizer)
 
     if arguments.use_agent_tokens or arguments.additional_vocabulary_tokens is not None:
         model = AutoPeftModelForCausalLM.from_pretrained(lora_dir)
@@ -140,8 +141,9 @@ def merge_base(arguments: MergeArguments, tokenizer, base_model, bnb_config) -> 
 
     print(f'Saving model to {model_dir}')
     print('')
-    if os.path.exists(model_dir):
+    if os.path.exists(model_dir) and arguments.overwrite_output:
         shutil.rmtree(model_dir)
+
     model.save_pretrained(model_dir)
     tokenizer.save_pretrained(model_dir)
     del model
